Version 1.7.6; Adding ESP8266 Gateway support. Be careful though, not all watchdog reset problems are ALWAYS solved. Especially when the ESP needs to reconnect to the network or the LamPI daemon there might be problems.

+++ libraries/LamPI/RemoteReceiver.h
@@ -9,7 +9,7 @@
 
 #include <Arduino.h>
 
-typedef void (*RemoteReceiverCallBack)(unsigned long, unsigned int);
+typedef void (*RemoteReceiverCallBack)(uint32_t, unsigned int);
 
 /**
 * See RemoteSwitch for introduction.
